feat: export MCP smoke test runner
Move the local MCP smoke corpus into a transport-neutral @githits/mcp/smoke-test subpath so remote MCP servers can reuse the same checks. Keep local unauthenticated stdio behavior in the script adapter and add package-boundary validation for the new public export.
