+++ scripts/validate-public-packages.ts
@@ -324,7 +324,7 @@ async function verifyMcpConsumer(
 
   await writeFile(
     join(appDirectory, "runtime-check.mjs"),
-    `import { createMcpServer, getMcpToolDescriptors } from "@githits/mcp";\nimport { CodeNavigationServiceImpl, PackageIntelligenceServiceImpl, GitHitsServiceImpl, createClientHeaderBuilder, createStaticTokenProvider, getApiUrl } from "@githits/mcp/client";\nif (typeof createMcpServer !== "function") throw new Error("missing createMcpServer");\nif (typeof CodeNavigationServiceImpl !== "function") throw new Error("missing CodeNavigationServiceImpl");\nif (typeof PackageIntelligenceServiceImpl !== "function") throw new Error("missing PackageIntelligenceServiceImpl");\nif (typeof GitHitsServiceImpl !== "function") throw new Error("missing GitHitsServiceImpl");\nif (typeof createStaticTokenProvider !== "function") throw new Error("missing createStaticTokenProvider");\nif (typeof createClientHeaderBuilder !== "function") throw new Error("missing createClientHeaderBuilder");\nif (typeof getApiUrl !== "function") throw new Error("missing getApiUrl");\nif (getMcpToolDescriptors().length === 0) throw new Error("missing descriptors");\ntry {\n  await import("@githits/mcp/internal");\n  throw new Error("internal export resolved");\n} catch (error) {\n  if (error instanceof Error && error.message === "internal export resolved") throw error;\n}\n`,
+    `import { createMcpServer, getMcpToolDescriptors } from "@githits/mcp";\nimport { CodeNavigationServiceImpl, PackageIntelligenceServiceImpl, GitHitsServiceImpl, createClientHeaderBuilder, createStaticTokenProvider, getApiUrl } from "@githits/mcp/client";\nimport { EXPECTED_MCP_TOOLS, runMcpSmoke } from "@githits/mcp/smoke-test";\nif (typeof createMcpServer !== "function") throw new Error("missing createMcpServer");\nif (typeof CodeNavigationServiceImpl !== "function") throw new Error("missing CodeNavigationServiceImpl");\nif (typeof PackageIntelligenceServiceImpl !== "function") throw new Error("missing PackageIntelligenceServiceImpl");\nif (typeof GitHitsServiceImpl !== "function") throw new Error("missing GitHitsServiceImpl");\nif (typeof createStaticTokenProvider !== "function") throw new Error("missing createStaticTokenProvider");\nif (typeof createClientHeaderBuilder !== "function") throw new Error("missing createClientHeaderBuilder");\nif (typeof getApiUrl !== "function") throw new Error("missing getApiUrl");\nif (typeof runMcpSmoke !== "function") throw new Error("missing runMcpSmoke");\nif (EXPECTED_MCP_TOOLS.length === 0) throw new Error("missing expected smoke tools");\nif (getMcpToolDescriptors().length === 0) throw new Error("missing descriptors");\ntry {\n  await import("@githits/mcp/internal");\n  throw new Error("internal export resolved");\n} catch (error) {\n  if (error instanceof Error && error.message === "internal export resolved") throw error;\n}\n`,
   );
   await runCommand(
     "node",
@@ -335,7 +335,7 @@ async function verifyMcpConsumer(
 
   await writeFile(
     join(appDirectory, "check.ts"),
-    `import { buildMcpInstructions, createMcpServer, getMcpToolDescriptors, type CreateMcpServerOptions, type McpToolServicesProvider } from "@githits/mcp";\nimport { CodeNavigationServiceImpl, GitHitsServiceImpl, PackageIntelligenceServiceImpl, createClientHeaderBuilder, createStaticTokenProvider, getApiUrl, getCodeNavigationUrl, type GitHitsService, type TokenProvider } from "@githits/mcp/client";\nconst provider: McpToolServicesProvider = () => { throw new Error("unused"); };\nconst options: CreateMcpServerOptions = { authAction: "Authenticate with the hosted GitHits MCP server.", metadata: { name: "consumer", version: "0.0.0" }, services: provider };\nconst tokenProvider: TokenProvider = createStaticTokenProvider("token");\nconst headers = createClientHeaderBuilder({ clientName: "remote-mcp", clientVersion: "0.0.0" });\nconst gitHitsService: GitHitsService = new GitHitsServiceImpl(getApiUrl(), "token", undefined, undefined, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nvoid new CodeNavigationServiceImpl(getCodeNavigationUrl(), tokenProvider, globalThis.fetch, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nvoid new PackageIntelligenceServiceImpl(getCodeNavigationUrl(), tokenProvider, globalThis.fetch, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nvoid gitHitsService;\nvoid createMcpServer(options);\nvoid buildMcpInstructions();\nif (getMcpToolDescriptors().length === 0) throw new Error("expected descriptors");\n`,
+    `import { buildMcpInstructions, createMcpServer, getMcpToolDescriptors, type CreateMcpServerOptions, type McpToolServicesProvider } from "@githits/mcp";\nimport { CodeNavigationServiceImpl, GitHitsServiceImpl, PackageIntelligenceServiceImpl, createClientHeaderBuilder, createStaticTokenProvider, getApiUrl, getCodeNavigationUrl, type GitHitsService, type TokenProvider } from "@githits/mcp/client";\nimport { EXPECTED_MCP_TOOLS, runMcpSmoke, type McpSmokeCaller } from "@githits/mcp/smoke-test";\nconst provider: McpToolServicesProvider = () => { throw new Error("unused"); };\nconst options: CreateMcpServerOptions = { authAction: "Authenticate with the hosted GitHits MCP server.", metadata: { name: "consumer", version: "0.0.0" }, services: provider };\nconst tokenProvider: TokenProvider = createStaticTokenProvider("token");\nconst headers = createClientHeaderBuilder({ clientName: "remote-mcp", clientVersion: "0.0.0" });\nconst gitHitsService: GitHitsService = new GitHitsServiceImpl(getApiUrl(), "token", undefined, undefined, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nconst caller: McpSmokeCaller = { listTools: async () => ({ tools: EXPECTED_MCP_TOOLS.map((name) => ({ name })) }), callTool: async () => ({ content: [{ type: "text", text: "ok" }] }) };\nvoid new CodeNavigationServiceImpl(getCodeNavigationUrl(), tokenProvider, globalThis.fetch, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nvoid new PackageIntelligenceServiceImpl(getCodeNavigationUrl(), tokenProvider, globalThis.fetch, { clientHeaders: headers, userAgent: "remote-mcp/0.0.0" });\nvoid gitHitsService;\nvoid createMcpServer(options);\nvoid buildMcpInstructions();\nvoid runMcpSmoke(caller, { includeLiveTools: false });\nif (getMcpToolDescriptors().length === 0) throw new Error("expected descriptors");\n`,
   );
   await writeFile(
     join(appDirectory, "tsconfig.json"),
