+++ tsconfig.json
@@ -14,6 +14,7 @@
     "paths": {
       "@githits/core-internal": ["./packages/core-internal/src/index.ts"],
       "@githits/mcp": ["./packages/mcp/src/index.ts"],
+      "@githits/mcp/smoke-test": ["./packages/mcp/src/smoke-test.ts"],
       "@githits/mcp/internal": ["./packages/mcp/src/internal.ts"]
     },
     "allowImportingTsExtensions": true,
